+++ frontend/src/__generated/index.css
@@ -6866,45 +6866,45 @@ body {
 .uok1x63 {
   color: var(--_1pyqka9d);
 }
-.gw96js0 {
+._1lhv06o0 {
   height: 40px;
 }
-.gw96js1 {
+._1lhv06o1 {
   height: calc(100% - 40px);
 }
-.gw96js2 {
+._1lhv06o2 {
   flex-grow: 1;
 }
-.gw96js3 {
+._1lhv06o3 {
   z-index: 0;
 }
-.gw96js4 {
+._1lhv06o4 {
   margin-bottom: 4px;
 }
-.gw96js5 {
+._1lhv06o5 {
   height: 28px;
 }
-.gw96js6 {
+._1lhv06o6 {
   width: 860px;
   height: 540px;
   margin: auto;
   z-index: 1;
   background-color: #ffffff;
 }
-.gw96js7 {
+._1lhv06o7 {
   width: 100%;
 }
-.gw96js8 {
+._1lhv06o8 {
   display: flex;
   justify-content: space-between;
   width: 100%;
   border-bottom: var(--_1pyqka9j) solid 1px;
   padding-bottom: 6px;
 }
-.gw96js9 {
+._1lhv06o9 {
   width: fit-content;
 }
-.gw96jsa {
+._1lhv06oa {
   background-color: var(--_1pyqka90);
   border-bottom-left-radius: 6px;
   border-bottom-right-radius: 6px;
