--- frontend/src/__generated/ve/pages/Graphing/GraphingEditor.css.js
@@ -1,25 +1,0 @@
-// src/pages/Graphing/GraphingEditor.css.ts
-var editGraphHeader = "gw96js0";
-var editGraphPanel = "gw96js1";
-var editGraphPreview = "gw96js2";
-var editorHeader = "gw96js8";
-var editorLabel = "gw96js4";
-var editorSelect = "gw96js9";
-var graphBackground = "gw96js3";
-var graphWrapper = "gw96js6";
-var input = "gw96js5";
-var sqlEditorWrapper = "gw96jsa";
-var tagSwitch = "gw96js7";
-export {
-  editGraphHeader,
-  editGraphPanel,
-  editGraphPreview,
-  editorHeader,
-  editorLabel,
-  editorSelect,
-  graphBackground,
-  graphWrapper,
-  input,
-  sqlEditorWrapper,
-  tagSwitch
-};
