+++ frontend/src/__generated/ve/pages/Graphing/GraphingEditor/GraphingEditor.css.js
@@ -1,0 +1,25 @@
+// src/pages/Graphing/GraphingEditor/GraphingEditor.css.ts
+var editGraphHeader = "_1lhv06o0";
+var editGraphPanel = "_1lhv06o1";
+var editGraphPreview = "_1lhv06o2";
+var editorHeader = "_1lhv06o8";
+var editorLabel = "_1lhv06o4";
+var editorSelect = "_1lhv06o9";
+var graphBackground = "_1lhv06o3";
+var graphWrapper = "_1lhv06o6";
+var input = "_1lhv06o5";
+var sqlEditorWrapper = "_1lhv06oa";
+var tagSwitch = "_1lhv06o7";
+export {
+  editGraphHeader,
+  editGraphPanel,
+  editGraphPreview,
+  editorHeader,
+  editorLabel,
+  editorSelect,
+  graphBackground,
+  graphWrapper,
+  input,
+  sqlEditorWrapper,
+  tagSwitch
+};
