+++ frontend/src/pages/Graphing/GraphingEditor/GraphingEditor.css.ts
